+++ src/main/ipc/handlers.ts
@@ -6,7 +6,5 @@ export function handle<K extends keyof IpcContract>(
   channel: K,
   handler: (...args: Parameters<IpcContract[K]>) => Promise<ReturnType<IpcContract[K]>>,
 ): void {
-  ipcMain.handle(channel, (_event, ...args) =>
-    handler(...(args as Parameters<IpcContract[K]>)),
-  );
+  ipcMain.handle(channel, (_event, ...args) => handler(...(args as Parameters<IpcContract[K]>)));
 }